+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Test data contract against dataframes / temporary views (#175)
+
 ## [0.10.6] - 2024-05-29
 
 ### Fixed